+++ src/Command.php
@@ -7,7 +7,7 @@ namespace mikehaertl\shellcommand;
  * This class represents a shell command.
  *
  * @author Michael Härtl <haertl.mike@gmail.com>
- * @version 1.0.4-dev
+ * @version 1.0.4
  * @license http://www.opensource.org/licenses/MIT
  */
 class Command