+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.0.4
+
+ * Add `$useExec` option to fix Windows issues (#3)
+
+## 1.0.3
+
+ * Add `getExecuted()` to find out execution status of the command
+
 ## 1.0.2
 
  * Add `$escape` parameter to `addArg()` to override escaping settings per call